+++ build-tools/Cargo.toml
@@ -4,4 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.0", features = ["derive"] }
+structopt = "0.3.23"
+
+aws-config = "0.51.0"
+aws-sdk-lambda = "0.21.0"
+aws-sdk-sts = "0.21.0"
+tokio = { version = "1", features = ["full"] }
